Changed validateIdToken to work with multiple keys
Previously the validateIdToken function would only take the first key
from the JWK set and attempt to use that o decode the jwt.  It will
now use all of them and either use the first success or fail.

I've also run hlint over the codebase and formatted it using brittany.